+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Schema changes: moved `full_name` field from table `userprofile` to table `user` for search improvement reasons.
 * Adding `messages_count`, `followers_count` and `following_count` to `profile_info` API endpoint (what I've done to 0.7.1 too).
+* Adding media URLs to messages in API.
+* Added `relationships_follow`, `relationships_unfollow`, `username_availability` and `edit_profile` endpoints to API.
+* Added `url` utility to model `Upload`.
 
 ## 0.7.1-dev